fix: move useTheme to hooks
