+++ src/components/sidebar/sidebar.tsx
@@ -2,9 +2,9 @@ import { FC } from "react";
 import Logo from "../ui/logo/logo";
 import NavButton from "../ui/button/nav-button/nav-button";
 import SettingButton from "@src/components/ui/button/setting-button/setting-button";
-import { useTheme } from "../theme-provider/theme-provider";
 import { useAppDispatch } from "@src/hooks/store-hooks";
 import { logoutThunk } from "@src/store/reducers/profileInfo/async-thunks";
+import { useTheme } from "@src/hooks/useTheme";
 
 const Sidebar: FC = () => {
   const dispatch = useAppDispatch();
